requirements.txt
autogen
python-dotenv
beautifulsoup4